Create Brand Identity Discovery Interview
For my brand identity exercise, I tried out this prompt in Gemini (Flash 2.0 Thinking) and found it really helpful. If feels a lot easier to be interviewed by the AI, instead of trying to answer questions on my own. You can probably use this in LM studio if you don't want to send answers to online LLM: